Update Partition size to 16 bits.
This will allow us to support up to 65k processors, which should be good for the forseeable future.

Needs to be tested a lot, especially if we can access that 256 core system again.

Closes #809

+++ src/core/registration.c
@@ -345,7 +345,7 @@ QuicRegistrationAcceptConnection(
         Connection->PartitionID += QUIC_MAX_THROUGHPUT_PARTITION_OFFSET;
     }
 
-    uint8_t Index =
+    uint16_t Index =
         Registration->NoPartitioning ? 0 : QuicPartitionIdGetIndex(Connection->PartitionID);
 
     //
@@ -366,7 +366,7 @@ QuicRegistrationQueueNewConnection(
     _In_ QUIC_CONNECTION* Connection
     )
 {
-    uint8_t Index =
+    uint16_t Index =
         Registration->NoPartitioning ? 0 : QuicPartitionIdGetIndex(Connection->PartitionID);
 
     //
