+++ src/inc/quic_datapath.h
@@ -167,7 +167,7 @@ typedef struct QUIC_RECV_DATAGRAM {
     //
     // The partition ID of the received datagram.
     //
-    uint8_t PartitionIndex;
+    uint16_t PartitionIndex;
 
     //
     // The Type of Service (ToS) field of the IPv4 header or Traffic Class field
